initial working version of streaming audio

+++ src/game_client.py
@@ -1,9 +1,11 @@
 import os
 import shutil
 from pathlib import Path
-
+import queue
 from dotenv import load_dotenv
+import threading 
 
+from signals import Signals
 import helper as h
 import tcp_file_transfer as tcp
 import audio_management as am
@@ -21,7 +23,7 @@ class GameClient:
                  server_ip=os.getenv("SERVER_IP"),
                  server_port=os.getenv("SERVER_PORT"),
                  record_time=RECORD_TIME,
-                 remove_temp=True):
+                 remove_temp=False):
         self.temp_dir = h.init_temp_storage(temp_dir_path)
 
         self.remove_temp = remove_temp
@@ -31,23 +33,39 @@ class GameClient:
         #  client setup
         self.tcpc = tcp.TCPClient(server_ip, server_port) # connects to server in init
         self.tcpc.connect_to_server()
+        
 
-    def main_loop_non_stream(self):
+    def main_loop(self):
+        output_queue = queue.Queue(maxsize=10)
+
         while True:
-            received_file_path = self.temp_dir / "received.wav"
+            signal = self.tcpc.receive_signal()
+            if signal == Signals.FILE_SENT:
+                received_file_path = self.temp_dir / "received.wav"
 
-            received_file_path = self.tcpc.receive_file(received_file_path)
-            if received_file_path is None: # handles bad inputs
-                break
+                received_file_path = self.tcpc.receive_file(received_file_path)
+                if received_file_path is None: # handles bad inputs
+                    break
+                am.play_audio(received_file_path)
+                os.remove(received_file_path)
+            elif signal == Signals.INIT_FT_STREAMED:
+                # separate thread for audio playback
+                audio_playback_thread = threading.Thread(target=am.play_audio_stream, 
+                                                    args=(output_queue,),
+                                                        daemon=True)
+                audio_playback_thread.start()
+                self.tcpc.receive_file_stream(output_queue, self.temp_dir / "playback", ".wav", file_name_tag="playback")
+                audio_playback_thread.join() # wait for playback to finish
+                # shutil.rmtree(self.temp_dir / "playback")
+            else:
+                # TODO: Handle other signals
+                raise NotImplementedError(f"Signal {Signals(signal).name} is not implemented")
 
-            am.play_audio(received_file_path)
 
             record_file_path = self.temp_dir / "recorded.wav"
-
             record_file_path = am.record_audio_by_time(record_file_path)
             self.tcpc.send_file(record_file_path)
 
-            os.remove(received_file_path)
             os.remove(record_file_path)
 
 
@@ -57,7 +75,7 @@ class GameClient:
 
 def main():
     game_client = GameClient()
-    game_client.main_loop_non_stream()
+    game_client.main_loop()
 
 
 if __name__ == '__main__':
